r14337: Try and quieten coverity #53 and #54. Make it obvious we're using -1 as a special size_t case by casting. Jeremy.

+++ source/smbd/msdfs.c
@@ -680,7 +680,7 @@ static int setup_ver2_dfs_referral(char *pathname, char **ppdata,
 		SSVAL(pdata,offset+18,uni_reqpathoffset2-offset);
 		/* copy referred path into current offset */
 		unilen = rpcstr_push(pdata+uni_curroffset, ref->alternate_path,
-				     -1, STR_UNICODE);
+				     (size_t)-1, STR_UNICODE);
 
 		SSVAL(pdata,offset+20,uni_curroffset-offset);
 
@@ -709,7 +709,7 @@ static int setup_ver3_dfs_referral(char *pathname, char **ppdata,
 	
 	DEBUG(10,("setting up version3 referral\n"));
 
-	reqpathlen = rpcstr_push(uni_reqpath, pathname, -1, STR_TERMINATE);
+	reqpathlen = rpcstr_push(uni_reqpath, pathname, (size_t)-1, STR_TERMINATE);
 	
 	if (DEBUGLVL(10)) {
 	    dump_data(0, (char *) uni_reqpath,reqpathlen);
